上传文件后保存文件数据到数据库，并且如果文件已存在就返回老的文件地址

+++ src/controllers/utilController.ts
@@ -4,6 +4,8 @@ import { successResponse } from "../models/Response";
 import path from "node:path";
 import fs from "node:fs";
 import { RunMode } from "../runMode";
+import prisma from "../plugins/prisma";
+import crypto from "node:crypto";
 
 type FileUploadBody = {
     file: Blob;
@@ -60,22 +62,45 @@ export const uploadFile = async (c: Context) => {
     }
     console.log('-----process.env.BASE_URL: ', file.type)
     const buffer = Buffer.from(await file.arrayBuffer());
+
+
+    const md5Value = crypto.createHash('md5').update(buffer).digest('hex');
+    const existed = await prisma.knowledgeFile.findFirst({
+      where: {
+        md5: md5Value
+      }
+    })
+    if (existed) {
+      return c.json(successResponse({ url: existed.filepath }, "文件已存在，返回已存储的文件 URL"));
+    }
     const objectKey = `${Date.now()}-${crypto.randomUUID().slice(0, 8)}-${sanitizeFilename(file.name)}`;
     const mode = c.var.mode as RunMode;
+    let url = ''
     try {
       if (mode === "local") {
         const filePath = path.join("static/uploaded", objectKey);
         fs.writeFileSync(filePath, buffer);
         const BASE_URL = process.env.BASE_URL!;
-        return c.json(successResponse({url: `${BASE_URL}/${filePath}`}, "文件上传成功"));
+        url = `${BASE_URL}/${filePath}`
       } else {
         await client.put(objectKey, buffer);
-        const url = client.signatureUrl(objectKey, {
+        url = client.signatureUrl(objectKey, {
           expires: OSS_SIGNED_URL_EXPIRES_SEC,
           method: "GET",
         });
-        return c.json(successResponse({ url }, "文件上传成功"));
+        console.log("Generated signed URL-----:", url);
+        // return c.json(successResponse({ url }, "文件上传成功"));
       }
+      await prisma.knowledgeFile.create({
+        data: {
+          filename: file.name,
+          filepath: url,
+          filetype: file.type,
+          filesize: file.size,
+          md5: md5Value
+        }
+      })
+      return c.json(successResponse({url}, "文件上传成功"));
     } catch (e) {
       console.error("OSS upload failed:", e);
       return c.json({ error: "Upload failed" }, 500);
